Move CollisionHandler to Game
Game should handle the Collisions, instead of GameState.

> organisation
